test(vetting): Wave 7 — rename gabor to test_gabor_regression
1-feature family rename (§6.4). git mv test_gabor.{h,cc} ->
test_gabor_regression.{h,cc}; test_gabor_truth.h fixture left as-is.
Updated the intra-.cc include, test_all.cc include, and the two
CMakeLists TEST_SRC entries. Registry current_test repointed.

VERIFIED: full gtest suite 696/696 — no tests dropped; registry rc=0.
