+++ tests/test_gabor_regression.h
