+++ tests/vetting/oracle_coverage.csv
@@ -307,7 +307,7 @@ dim,feature,family,status,oracle,agreement,config_recipe,tolerance,current_test,
 2D,NGTDM_COMPLEXITY,ngtdm,vetted,pyradiomics,agreed,"Mode-specific 2D texture row. PyRadiomics/MIRP comparison is tied to verifier settings such as force2D/by_slice, binWidth/binCount or no discretization, symmetry, and aggregation.",,test_3d_ngtdm.h;test_3d_ngtdm_pyradiomics.h;test_ngtdm_ibsi.h;test_ngtdm_regression.h,test_ngtdm_pyradiomics.h,,,tracker,test_3d_ngtdm.h ORPHANED (not #included; never run) - left per decision A
 2D,NGTDM_STRENGTH,ngtdm,vetted,pyradiomics,agreed,"Mode-specific 2D texture row. PyRadiomics/MIRP comparison is tied to verifier settings such as force2D/by_slice, binWidth/binCount or no discretization, symmetry, and aggregation.",,test_3d_ngtdm.h;test_3d_ngtdm_pyradiomics.h;test_ngtdm_ibsi.h;test_ngtdm_regression.h,test_ngtdm_pyradiomics.h,,,tracker,test_3d_ngtdm.h ORPHANED (not #included; never run) - left per decision A
 2D,FRAC_AT_D,radial,regression,,na,"Not mode-specific. Any disagreement is a definition, coordinate-frame, spacing, aggregation, or implementation issue rather than Nyxus IBSI/radiomics mode.",,test_intensity_histogram_regression.h,test_intensity_histogram_regression.h,cellprofiler:MeasureObjectIntensityDistribution (RadialDistribution_*),oracle-identified,tracker,
-2D,GABOR,gabor,regression,,na,"Not mode-specific. Any disagreement is a definition, coordinate-frame, spacing, aggregation, or implementation issue rather than Nyxus IBSI/radiomics mode.",,test_gabor.cc;test_nyxus.py,test_gabor_regression.h,skimage/MATLAB gabor kernel (no scalar oracle),no-mainstream-oracle,tracker,
+2D,GABOR,gabor,regression,,na,"Not mode-specific. Any disagreement is a definition, coordinate-frame, spacing, aggregation, or implementation issue rather than Nyxus IBSI/radiomics mode.",,test_gabor_regression.cc;test_nyxus.py,test_gabor_regression.h,skimage/MATLAB gabor kernel (no scalar oracle),no-mainstream-oracle,tracker,
 2D,MEAN_FRAC,radial,regression,,na,"Not mode-specific. Any disagreement is a definition, coordinate-frame, spacing, aggregation, or implementation issue rather than Nyxus IBSI/radiomics mode.",,test_intensity_histogram_regression.h,test_intensity_histogram_regression.h,cellprofiler:MeasureObjectIntensityDistribution (RadialDistribution_*),oracle-identified,tracker,
 2D,RADIAL_CV,radial,regression,,na,"Not mode-specific. Any disagreement is a definition, coordinate-frame, spacing, aggregation, or implementation issue rather than Nyxus IBSI/radiomics mode.",,test_intensity_histogram_regression.h,test_intensity_histogram_regression.h,cellprofiler:MeasureObjectIntensityDistribution (RadialDistribution_*),oracle-identified,tracker,
 2D,ZERNIKE2D,zernike,regression,,benign,"Not mode-specific. Any disagreement is a definition, coordinate-frame, spacing, aggregation, or implementation issue rather than Nyxus IBSI/radiomics mode.",,test_zernike_regression.h,test_zernike_regression.h,"none (mahotas is niche, not an accepted oracle) -> analytic/regression",no-mainstream-oracle,tracker,
